requirements.txt
numpy>=1.24.0
Pillow>=10.0.0
opencv-python>=4.8.0
scipy>=1.11.0
scikit-image>=0.21.0
pywavelets>=1.4.0
matplotlib>=3.7.0
